Enable inbound SIP calls
Enable inbound SIP calls.  There are a few steps to this.

1) iptables config.  Open UDP and TCP port 5060 for SIP, as well as
UDP ports 10000-20000 for RTP.

2) Add a custom sip.conf which makes chan_sip listen on all address, including
IPv4 and IPv6.  Also enable unauthenticated inbound calls and send them to the
'public' dialplan context.

3) Create the dialplan.  Right now it just plays a sound prompt called 'spam'.
You'll have to call in to find out what it says.  Note that this required
installing the extra sounds.  There's a bunch of good stuff in there that
may be handy, other than just 'spam'.

Change-Id: I6b62511317603eedf9280b55a00ba5cee0611b62