Add explicit ordering for module reload execs
The new resource ordering algorithm in puppet 4 makes it so that it is
possible that the module reload execs occur before the module file
exists. Most of the time, this does not matter because we use a dummy
command in the exec, but for real reloads this could cause the exec to
fail. This patch adds explicit ordering to all the module reload execs,
even the dummy ones for consistency.

Change-Id: Ic892983476c7da60e4801779dfecc50587a2404d