Ensure asterisk refresh happens last
The new resource ordering algorithm in puppet 4 may cause the asterisk
service to be refreshed before the module directories are in place,
which then causes the module reload exec to fail. This patch adds an
explicit ordering to the customdir defined type to ensure that the
module directory exists before the asterisk service is refreshed.

Change-Id: Ic892983476c7da60e4801779dfecc50587a2404d